public check_rotations method and test.
