+++ src/aspire/utils/__init__.py
@@ -1,6 +1,7 @@
 from .types import complex_type, real_type, utest_tolerance  # isort:skip
 from .coor_trans import (  # isort:skip
     common_line_from_rots,
+    check_rotations,
     crop_pad_2d,
     crop_pad_3d,
     get_aligned_rotations,
